+++ libbpf-rs/src/print.rs
@@ -1,7 +1,7 @@
 use crate::libbpf_sys;
 use lazy_static::lazy_static;
+use std::io;
 use std::io::Write;
-use std::io::{self};
 use std::os::raw::c_char;
 use std::sync::Mutex;