+++ libbpf-rs/src/program.rs
@@ -17,6 +17,7 @@ use strum_macros::Display;
 
 use crate::libbpf_sys;
 use crate::util;
+use crate::AsRawLibbpf;
 use crate::Error;
 use crate::Link;
 use crate::Result;
@@ -228,6 +229,15 @@ impl OpenProgram {
         let count = self.insn_cnt();
         let ptr = unsafe { libbpf_sys::bpf_program__insns(self.ptr.as_ptr()) };
         unsafe { std::slice::from_raw_parts(ptr, count) }
+    }
+}
+
+impl AsRawLibbpf for Program {
+    type LibbpfType = libbpf_sys::bpf_program;
+
+    /// Retrieve the underlying [`libbpf_sys::bpf_program`].
+    fn as_libbpf_object(&self) -> NonNull<Self::LibbpfType> {
+        self.ptr
     }
 }